+++ app/Http/Controllers/User/AlbumsController.php
@@ -6,12 +6,10 @@ use Illuminate\Routing\Router;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 
-use App\Models\Categories;
 use App\Models\Albums;
-use App\Models\Tags;
-use App\Models\Images;
 
 use Auth;
+use Agent;
 use Roles;
 use Setting;
 use Viewer;
@@ -19,118 +17,67 @@ use Cache;
 
 class AlbumsController extends Controller
 {
-    protected $categories;    
     protected $albums;
-    protected $tags;    
-    protected $images;
-    
-    public function __construct(Categories $categories, Albums $albums, Tags $tags, Images $images) {
-        $this->categories = $categories;
-        $this->albums     = $albums;
-        $this->tags       = $tags;
-        $this->images     = $images;       
+
+    public function __construct(Albums $albums) {
+        $this->albums  = $albums;
     }
     
-    public function getShow(Router $router) {
-                
-        if(Roles::is('admin')) {
+    /*
+     * Вывод фотографий альбома
+     */
+    public function getPage(Router $router, Request $request) {
         
-            $countShowAlbums = Cache::remember('all.showadmin.albums', Setting::get('cache_ttl'), function() {
-                return $this->albums->where('images_id', '!=', '0')->count();
-            });
-            
-            $CacheKey = 'Admin.';
-            
-        } else {
-            
-            $countShowAlbums = Cache::remember('all.public.albums', Setting::get('cache_ttl'), function() {
-                return $this->albums->where('permission', 'All')->where('images_id', '!=', '0')->count();
-            });
-            
-            $CacheKey = 'User.';
-        }
+        $url = $router->input('url');
         
-        if($countShowAlbums == 0)
-            return Viewer::get('errors.404');         
+        $thisAlbum = Cache::remember('albums.show.' . $url, Setting::get('cache_ttl'), function() use ($url){
+            return $this->albums->where('url', $url)->first();
+        });
         
-        $CacheKey .= 'Cache.Albums.';
-        $CacheKey .= (!empty($router->input('option'))) ? $router->input('option') : "Base" ;
-        $CacheKey .= ($router->input('url')) ? "." . $router->input('url') : ".Null" ;
+        if(!$thisAlbum)
+            return Viewer::get('errors.404');
+        
+        if (Agent::isMobile() and !Auth::check())
+            $CacheKey = $url . '.Mobile' . $request->input('page');
+        elseif(Agent::isMobile() and Roles::is('admin'))
+            $CacheKey = $url . '.AdminMobile' . $request->input('page');
+        elseif(!Agent::isMobile() and Roles::is('admin'))
+            $CacheKey = $url . '.Admin' . $request->input('page');        
+        else
+            $CacheKey = $url . $request->input('page');
         
         if (Cache::has($CacheKey)) {
-            
             $resultData = Cache::get($CacheKey);
-            
         } else {
-            
-            if($router->input('option') == "tag" and !empty($router->input('url'))) {
-                
-                $tags = $this->tags->select('id')->where('name', urldecode($router->input('url')))->firstOrFail();
-                $albums = $tags->albums;
-                
-                if(!Roles::is('admin'))
-                    $albums = $albums->where('permission', 'All');
-            
-                $albums = $albums->where('images_id', '!=', '0');                
-                
-            } elseif ($router->input('option') == "category" and !empty($router->input('url'))) {
-                
-                $category = $this->categories->select('id')->where('name', urldecode($router->input('url')))->firstOrFail();
-                $albums = $category->albums;
-                
-                if(!Roles::is('admin'))
-                    $albums = $albums->where('permission', 'All');
-            
-                $albums = $albums->where('images_id', '!=', '0');
-                
-            } elseif ($router->input('option') == "year" and !empty($router->input('url'))){
-                
-                $albums = $this->albums->latest('year');
-                $albums = $albums->where('year', $router->input('url'));
-                
-                if(!Roles::is('admin'))
-                    $albums = $albums->where('permission', 'All');
-            
-                $albums = $albums->where('images_id', '!=', '0');                
-                $albums = $albums->get();
-                
-            } else {
-                
-                $albums = $this->albums->latest('year');
-                
-                if(!Roles::is('admin'))
-                    $albums = $albums->where('permission', 'All');
-            
-                $albums = $albums->where('images_id', '!=', '0');
-                $albums = $albums->get();
-                
-            }
-
+        
             $thumbs_width  = Setting::get('thumbs_width');
             $thumbs_height = Setting::get('thumbs_height');
             $thumbs_dir    = Setting::get('thumbs_dir');
             
-            if(Roles::is('admin'))
-                $tags = $this->tags->whereHas('albums', function ($query) {
-                    $query->where('images_id', '!=', '0');
-                })->get();
+            if(!Agent::isMobile() and Roles::is('admin'))
+                $show_admin_panel = 1;
             else
-                $tags = $this->tags->whereHas('albums', function ($query) {
-                    $query->where('permission', 'All')->where('images_id', '!=', '0');
-                })->get();
+                $show_admin_panel = 0;             
             
+            $listImages = $thisAlbum
+                    ->images()
+                    ->paginate(Setting::get('count_images'));
+
             $resultData = compact(
-                'albums', 
-                'tags', 
-                'thumbs_width', 
-                'thumbs_height', 
-                'thumbs_dir'
+                'thumbs_width',
+                'thumbs_height',
+                'show_admin_panel',
+                'thisAlbum',
+                'listImages'
             );
             
             Cache::add($CacheKey, $resultData, Setting::get('cache_ttl'));
         }
         
+        if($request->ajax())
+            return Viewer::get('user.album.page', $resultData);
+        
         return Viewer::get('user.album.index', $resultData);
-
+        
     }
 }
