+++ app/Models/Images.php
@@ -13,7 +13,7 @@ class Images extends Model
     protected $dates = ['deleted_at'];
     
     protected $fillable = [
-        'name', 'size', 'albums_id', 'users_id', 'is_rebuild', 'is_thumbs', 'is_modile'
+        'name', 'size', 'height', 'width', 'albums_id', 'users_id', 'is_rebuild'
     ];
     
     public function album() {
@@ -70,21 +70,21 @@ class Images extends Model
     
     public function destroyImage($id){
         
-        $full_image   = \Helper::getFullPathImage($id);
-        $mobile_image = \Helper::getFullPathMobileImage($id);
-        $thumb_image  = \Helper::getFullPathThumbImage($id);
-        
-        if (\File::exists($full_image))
-            \File::delete($full_image);
-        
-        if (\File::exists($mobile_image))
-            \File::delete($mobile_image);
-        
-        if (\File::exists($thumb_image))
-            \File::delete($thumb_image);        
-        
-        $image = $this->withTrashed()->findOrFail($id);
-        $image->forceDelete();
+//        $full_image   = \Helper::getFullPathImage($id);
+//        $mobile_image = \Helper::getFullPathMobileImage($id);
+//        $thumb_image  = \Helper::getFullPathThumbImage($id);
+//        
+//        if (\File::exists($full_image))
+//            \File::delete($full_image);
+//        
+//        if (\File::exists($mobile_image))
+//            \File::delete($mobile_image);
+//        
+//        if (\File::exists($thumb_image))
+//            \File::delete($thumb_image);        
+//        
+//        $image = $this->withTrashed()->findOrFail($id);
+//        $image->forceDelete();
         
     }
 }
