+++ resources/views/default/user/album/page.blade.php
