+++ routes/web.php
@@ -22,7 +22,7 @@ Auth::routes();
 // Публичные страницы, доступные без авторизации
 Route::get('/', [
     'as'    => 'home', 
-    'uses'  => 'User\AlbumsController@getShow'
+    'uses'  => 'User\GalleryController@getShow'
 ]);
 
 Route::get('/no_access', [
@@ -32,12 +32,12 @@ Route::get('/no_access', [
 
 Route::get('/images/{option}/{url}/{name}', [
     'as'    => 'show-image', 
-    'uses'  => 'User\ImagesController@image'
+    'uses'  => 'User\ImagesController@getImage'
 ]);
 
 Route::get('/by/{option}/{url}', [
     'as'    => 'album-showBy', 
-    'uses'  => 'User\AlbumsController@getShow'    
+    'uses'  => 'User\GalleryController@getShow'    
 ])->where(['option', '[A-Za-z0-9]+', 'id' => '[А-Яа-яA-Za-z0-9_-]+']);
 
 Route::post('/ulogin', 'User\UloginController@login');
@@ -85,5 +85,5 @@ Route::group(['middleware' => 'auth'], function () {
 
 Route::get('/{url}', [
     'as'    => 'gallery-show', 
-    'uses'  => 'User\ImagesController@getPage'    
+    'uses'  => 'User\AlbumsController@getPage'    
 ])->where('url', '[А-Яа-яA-Za-z0-9_-]+');
